+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## [Unreleased] - yyyy-mm-dd
 
+## [0.2.0](https://code.sbb.ch/projects/KI_ROKAS/repos/journey-maps-client/compare/commits?targetBranch=refs%2Ftags%2F0.1.1&sourceBranch=refs%2Ftags%2F0.2.0&targetRepoId=46287) - 2021-01-04
+
 ### IMPORTANT
-You need a new API key when upgrading to this version. Please sign up for `Journey-Maps-Tiles` [here](https://developer.sbb.ch/apis/journey-maps-tiles/information).
+
+You need a new API key when upgrading to this version. Please sign up
+for `Journey-Maps-Tiles` [here](https://developer.sbb.ch/apis/journey-maps-tiles/information).
+
 ### Added
 
 * We added a search bar to filter the markers on the map. The search bar can be hidden using the input